Move review task machine into a folder
