+++ src/components/ClaimReview/ClaimReviewHeader.tsx
@@ -5,13 +5,13 @@ import React, { useContext, useEffect, useState } from "react";
 
 import ClaimReviewApi from "../../api/claimReviewApi";
 import { ReviewTaskMachineContext } from "../../Context/ReviewTaskMachineProvider";
-import { Roles } from "../../machine/enums";
+import { Roles } from "../../machines/reviewTask/enums";
 import {
     crossCheckingSelector,
     publishedSelector,
     reviewDataSelector,
     reviewNotStartedSelector,
-} from "../../machine/selectors";
+} from "../../machines/reviewTask/selectors";
 import { useAppSelector } from "../../store/store";
 import colors from "../../styles/colors";
 import AletheiaAlert from "../AletheiaAlert";
