+++ src/machines/reviewTask/getNextForm.ts
@@ -1,9 +1,9 @@
 import { ReviewTaskEvents, ReviewTaskStates } from "./enums";
-import assignedForm from "../components/ClaimReview/form/fieldLists/assignedForm";
-import reportedForm from "../components/ClaimReview/form/fieldLists/reportedForm";
-import unassignedForm from "../components/ClaimReview/form/fieldLists/unassignedForm";
-import rejectedForm from "../components/ClaimReview/form/fieldLists/rejectedForm";
-import summarizedForm from "../components/ClaimReview/form/fieldLists/summarizedForm";
+import assignedForm from "../../components/ClaimReview/form/fieldLists/assignedForm";
+import reportedForm from "../../components/ClaimReview/form/fieldLists/reportedForm";
+import unassignedForm from "../../components/ClaimReview/form/fieldLists/unassignedForm";
+import rejectedForm from "../../components/ClaimReview/form/fieldLists/rejectedForm";
+import summarizedForm from "../../components/ClaimReview/form/fieldLists/summarizedForm";
 
 const getNextForm = (param: ReviewTaskEvents | ReviewTaskStates) => {
     const formMap = {
